Move MessageBirdApi to helper packages.
