+++ app/helper/MessageBirdApi.scala
@@ -1,4 +1,4 @@
-package services
+package helper
 
 import helper.Time
 import java.nio.charset.StandardCharsets.UTF_8
